pyproject.toml
[tool.poetry]
name = "fantasy-f1-discord-bot"
version = "0.1.0"
description = ""
authors = ["Rasika Amaratissa <rasika.am@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
"discord.py" = "^1.7.3"
table2ascii = "^0.2.0"
requests = "^2.27.1"
python-dotenv = "^0.20.0"
emoji-country-flag = "^1.3.0"

[tool.poetry.dev-dependencies]
black = "^22.3.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
